Get rid of the hardcoded external process simulator name
Now, when building and running tests through the general Makefile, the name of the simulation process and the value of the variable with the process name for testing will be the same.

If the tests and the external process simulator are built not through the general Makefile, the resulting .exe files will also successfully interact with each other, provided that the external process simulator is not renamed by hand.