Implement constructor that takes process handle
This constructor takes a process handle and assigns it to '_handle' field.

Also, modify the constructor that takes a process id so that it calls OpenProcess function and delegates the result to the constructor that takes a process handle.

Remove '_process_id' field, since it is only needed as an intermediate step in the chain of obtaining a process handle by process name.

Close process handle in destructor.